[SQUASH] Merge tag 'LA.UM.9.14.r1-20700-LAHAINA.QSSI13.0' into staging/lineage-20.0_merge-LA.UM.9.14.r1-20700-LAHAINA.QSSI13.0
"LA.UM.9.14.r1-20700-LAHAINA.QSSI13.0"

# By Krishna Kishor Jha (2) and others
# Via Gerrit - the friendly Code Review server (5) and Linux Build Service Account (1)
* tag 'LA.UM.9.14.r1-20700-LAHAINA.QSSI13.0':
  audio-hal: add missing modules in audio_effects.xml
  hal: changing mtmx coeffs variable type to uint32_t
  audio-hal: automotive LL capture usecase changes blocking with flag
  config: added audio effects config for elite architecture
  config: added audio effect libraries and helpers.
  audio-hal: Fix build error with VA and AOSP tinycompress

+++ hal/platform_api.h
@@ -149,7 +149,7 @@ struct audio_custom_mtmx_params_info {
 struct audio_custom_mtmx_params {
     struct listnode list;
     struct audio_custom_mtmx_params_info info;
-    uint64_t coeffs[0];
+    uint32_t coeffs[0];
 };
 
 struct audio_custom_mtmx_in_params_info {
